+++ header/Server/Server.h
@@ -23,7 +23,7 @@
 #include <queue>
 #include <fstream>
 
-#include "Client.h"
+#include "Server/Client/Client.h"
