+++ source/Server/Client/Client.cpp
@@ -2,7 +2,7 @@
 // Created by matvey on 05.04.23.
 //
 
-#include "Client.h"
+#include "Server/Client/Client.h"
 
 void Client::update_ping() {
     last_ping = boost::posix_time::microsec_clock::local_time();
